+++ builds/4chan-X.meta.js
@@ -5,7 +5,7 @@
 // @minFFVer     26
 // @namespace    4chan-X
 // @description  Cross-browser userscript for maximum lurking on 4chan.
-// @license      MIT; https://github.com/seaweedchan/4chan-x/blob/master/LICENSE 
+// @license      MIT; https://github.com/Spittie/4chan-x/blob/master/LICENSE 
 // @match        *://boards.4chan.org/*
 // @match        *://sys.4chan.org/*
 // @match        *://a.4cdn.org/*
@@ -15,7 +15,7 @@
 // @grant        GM_deleteValue
 // @grant        GM_openInTab
 // @run-at       document-start
-// @updateURL 	 https://github.com/seaweedchan/4chan-x/raw/stable/builds/4chan-X.meta.js
-// @downloadURL  https://github.com/seaweedchan/4chan-x/raw/stable/builds/4chan-X.user.js
+// @updateURL 	 https://github.com/Spittie/4chan-x/raw/stable/builds/4chan-X.meta.js
+// @downloadURL  https://github.com/Spittie/4chan-x/raw/stable/builds/4chan-X.user.js
 // @icon         data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADAAAAAwAgMAAAAqbBEUAAAACVBMVEUAAGcAAABmzDNZt9VtAAAAAXRSTlMAQObYZgAAAHFJREFUKFOt0LENACEIBdBv4Qju4wgWanEj3D6OcIVMKaitYHEU/jwTCQj8W75kiVCSBvdQ5/AvfVHBin11BgdRq3ysBgfwBDRrj3MCIA+oAQaku/Q1cNctrAmyDl577tOThYt/Y1RBM4DgOHzM0HFTAyLukH/cmRnqAAAAAElFTkSuQmCC
 // ==/UserScript==