+++ builds/4chan-X.user.js
@@ -6,7 +6,7 @@
 // @minFFVer     26
 // @namespace    4chan-X
 // @description  Cross-browser userscript for maximum lurking on 4chan.
-// @license      MIT; https://github.com/seaweedchan/4chan-x/blob/master/LICENSE 
+// @license      MIT; https://github.com/Spittie/4chan-x/blob/master/LICENSE 
 // @match        *://boards.4chan.org/*
 // @match        *://sys.4chan.org/*
 // @match        *://a.4cdn.org/*
@@ -16,8 +16,8 @@
 // @grant        GM_deleteValue
 // @grant        GM_openInTab
 // @run-at       document-start
-// @updateURL 	 https://github.com/seaweedchan/4chan-x/raw/stable/builds/4chan-X.meta.js
-// @downloadURL  https://github.com/seaweedchan/4chan-x/raw/stable/builds/4chan-X.user.js
+// @updateURL 	 https://github.com/Spittie/4chan-x/raw/stable/builds/4chan-X.meta.js
+// @downloadURL  https://github.com/Spittie/4chan-x/raw/stable/builds/4chan-X.user.js
 // @icon         data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADAAAAAwAgMAAAAqbBEUAAAACVBMVEUAAGcAAABmzDNZt9VtAAAAAXRSTlMAQObYZgAAAHFJREFUKFOt0LENACEIBdBv4Qju4wgWanEj3D6OcIVMKaitYHEU/jwTCQj8W75kiVCSBvdQ5/AvfVHBin11BgdRq3ysBgfwBDRrj3MCIA+oAQaku/Q1cNctrAmyDl577tOThYt/Y1RBM4DgOHzM0HFTAyLukH/cmRnqAAAAAElFTkSuQmCC
 // ==/UserScript==
 
@@ -25,7 +25,7 @@
 * 4chan X - Version 1.3.3 - 2014-02-09
 *
 * Licensed under the MIT license.
-* https://github.com/seaweedchan/4chan-x/blob/master/LICENSE
+* https://github.com/Spittie/4chan-x/blob/master/LICENSE
 *
 * Appchan X Copyright © 2013-2013 Zixaphir <zixaphirmoxphar@gmail.com>
 * http://zixaphir.github.io/appchan-x/ 
@@ -6132,7 +6132,8 @@
         title: 'Verification',
         placeholder: 'Focus to load reCAPTCHA',
         autocomplete: 'off',
-        spellcheck: false
+        spellcheck: false,
+        tabIndex: 55
       });
       this.nodes = {
         img: imgContainer.firstChild,
@@ -9914,7 +9915,7 @@
           software: "foolfuuka"
         }
       },
-      "Install Gentoo": {
+      "Install Gentoo (.net)": {
         boards: ["diy", "g", "sci"],
         files: [],
         data: {
@@ -9922,6 +9923,16 @@
           http: false,
           https: true,
           software: "fuuka"
+        }
+      },
+      "Install Gentoo (.com)": {
+        boards: ["t", "g"],
+        files: ["t", "g"],
+        data: {
+          domain: "archive.installgentoo.com/",
+          http: true,
+          https: true,
+          software: "foolfuuka"
         }
       },
       "Rebecca Black Tech": {
@@ -11713,7 +11724,7 @@
           if (previous === g.VERSION) {
             return;
           }
-          changelog = 'https://github.com/seaweedchan/4chan-x/blob/master/CHANGELOG.md';
+          changelog = 'https://github.com/Spittie/4chan-x/blob/master/CHANGELOG.md';
           el = $.el('span', {
             innerHTML: "4chan X has been updated to <a href='" + changelog + "' target=_blank>version " + g.VERSION + "</a>."
           });
@@ -11748,7 +11759,7 @@
         return;
       }
       $.event('CloseMenu');
-      html = "<nav><div class=sections-list></div><p class='imp-exp-result warning'></p><div class=credits><a class=export>Export</a> |<a class=import>Import</a> |<input type=file style='display: none;'><a href='http://seaweedchan.github.io/4chan-x/' target=_blank>4chan X</a> |<a href='https://github.com/seaweedchan/4chan-x/blob/master/CHANGELOG.md' target=_blank>" + g.VERSION + "</a> |<a href='https://github.com/seaweedchan/4chan-x/blob/master/README.md#reporting-bugs-and-suggestions' target=_blank>Issues</a> |<a href=javascript:; class=close title=Close>×</a></div></nav><div class=section-container><section></section></div>";
+      html = "<nav><div class=sections-list></div><p class='imp-exp-result warning'></p><div class=credits><a class=export>Export</a> |<a class=import>Import</a> |<input type=file style='display: none;'><a href='http://seaweedchan.github.io/4chan-x/' target=_blank>4chan X</a> |<a href='https://github.com/Spittie/4chan-x/blob/master/CHANGELOG.md' target=_blank>" + g.VERSION + "</a> |<a href='https://github.com/Spittie/4chan-x/blob/master/README.md#reporting-bugs-and-suggestions' target=_blank>Issues</a> |<a href=javascript:; class=close title=Close>×</a></div></nav><div class=section-container><section></section></div>";
       Settings.overlay = overlay = $.el('div', {
         id: 'overlay'
       });